feat: expand template search and docs
